Remove use of oteltest in otelsarama
