+++ .github/dependabot.yml
@@ -307,6 +307,16 @@ updates:
       day: "sunday"
   -
     package-ecosystem: "gomod"
+    directory: "/instrumentation/github.com/Shopify/sarama/otelsarama/test"
+    labels:
+      - dependencies
+      - go
+      - "Skip Changelog"
+    schedule:
+      interval: "weekly"
+      day: "sunday"
+  -
+    package-ecosystem: "gomod"
     directory: "/instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo"
     labels:
       - dependencies
