+++ instrumentation/github.com/Shopify/sarama/otelsarama/go.mod
@@ -9,6 +9,5 @@ require (
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/contrib v0.22.0
 	go.opentelemetry.io/otel v1.0.0-RC2
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 )
