+++ instrumentation/github.com/Shopify/sarama/otelsarama/option.go
@@ -17,16 +17,11 @@ package otelsarama
 import (
 	"go.opentelemetry.io/contrib"
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 )
 
-const (
-	defaultTracerName = "go.opentelemetry.io/contrib/instrumentation/github.com/Shopify/sarama/otelsarama"
-
-	kafkaPartitionKey = attribute.Key("messaging.kafka.partition")
-)
+const defaultTracerName = "go.opentelemetry.io/contrib/instrumentation/github.com/Shopify/sarama/otelsarama"
 
 type config struct {
 	TracerProvider trace.TracerProvider
